refactor(rest-api): move release adder codes to correct files
