+++ rest-api/src/file_auth/mod.rs
@@ -2,4 +2,6 @@ pub mod actors;
 pub mod forges;
 pub mod forges_types;
 pub mod github;
+pub mod github_release;
 pub mod gitlab;
+pub mod releases;
